proved kernelc tree height and size examples
