+++ samples/kernelc/tests/config.xml
@@ -21,6 +21,8 @@
 
     <include file="list/config.xml" />
 
+    <include file="tree/config.xml" />
+
     <!--<include file="binary_search_tree//config.xml" more-results="." />-->
     <include file="binary_search_tree/config.xml" />
 </tests>
