+++ samples/kernelc/tests/tree/height.c.out
@@ -1,0 +1,1 @@
+true
